requirements.txt
requests
flask_cors
pandas